requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
langchain
langchain-openai
chromadb
pypdf
streamlit
requests